pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jarvis"
version = "0.1.0"
description = "Intelligent Traffic Intelligence Platform"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "numpy>=1.24",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "pyyaml>=6.0",
    "loguru>=0.7",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "alembic>=1.12",
    "pgvector>=0.2",
    "python-dotenv>=1.0",
    "tqdm>=4.66",
]

[project.optional-dependencies]
ml = [
    "torch>=2.1",
    "torchvision>=0.16",
    "ultralytics>=8.1",
    "transformers>=4.51",
    "opencv-python>=4.8",
    "Pillow>=10.0",
    "albumentations>=1.3",
    "scikit-learn>=1.3",
    "boxmot>=11.0",
    "sentence-transformers>=2.2",
    "peft>=0.7",
    "accelerate>=0.25",
    "bitsandbytes>=0.42",
]
api = [
    "fastapi>=0.104",
    "uvicorn>=0.24",
]
dashboard = [
    "streamlit>=1.29",
    "folium>=0.15",
    "plotly>=5.18",
]
cache = [
    "redis>=5.0",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "pre-commit>=3.5",
]
all = [
    "jarvis[ml,api,dashboard,cache,dev]",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: tests requiring a running database",
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
fail_under = 40
show_missing = true
skip_empty = true
